requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
fake-useragent==0.1.11
filelock==3.3.1
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
gdown==4.2.0
greenlet==1.1.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
lxml==4.6.3
MarkupSafe==2.0.1
multidict==5.1.0
numpy==1.21.2
opencv-python==4.5.4.58
pandas==1.3.3
Pillow==8.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
SQLAlchemy==1.4.25
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
yarl==1.6.3
